refactor: rename web to fetch_json
