+++ packages/providers/src/index.ts
@@ -1,4 +1,4 @@
 export { exponentialBackoff } from './exponential-backoff';
 export { JsonRpcProvider } from './json-rpc-provider';
 export { Provider } from './provider';
-export { fetchJson } from './web';
+export { fetchJson } from './fetch_json';
